+++ bin/miden-cli/src/commands/export.rs
@@ -3,6 +3,7 @@ use std::{fs::File, io::Write, path::PathBuf};
 use miden_client::{
     Client, Word,
     account::{Account, AccountFile},
+    auth::TransactionAuthenticator,
     store::NoteExportType,
     transaction::AccountInterface,
     utils::Serializable,
@@ -56,7 +57,11 @@ impl From<&ExportType> for NoteExportType {
 }
 
 impl ExportCmd {
-    pub async fn execute(&self, mut client: Client, keystore: CliKeyStore) -> Result<(), CliError> {
+    pub async fn execute<AUTH: TransactionAuthenticator>(
+        &self,
+        mut client: Client<AUTH>,
+        keystore: CliKeyStore,
+    ) -> Result<(), CliError> {
         if self.account {
             export_account(&client, &keystore, self.id.as_str(), self.filename.clone()).await?;
         } else if let Some(export_type) = &self.export_type {
@@ -73,8 +78,8 @@ impl ExportCmd {
 // EXPORT ACCOUNT
 // ================================================================================================
 
-async fn export_account(
-    client: &Client,
+async fn export_account<AUTH>(
+    client: &Client<AUTH>,
     keystore: &CliKeyStore,
     account_id: &str,
     filename: Option<PathBuf>,
@@ -120,8 +125,8 @@ async fn export_account(
 // EXPORT NOTE
 // ================================================================================================
 
-async fn export_note(
-    client: &mut Client,
+async fn export_note<AUTH: TransactionAuthenticator>(
+    client: &mut Client<AUTH>,
     note_id: &str,
     filename: Option<PathBuf>,
     export_type: &ExportType,