+++ bin/miden-cli/src/commands/tags.rs
@@ -23,7 +23,7 @@ pub struct TagsCmd {
 }
 
 impl TagsCmd {
-    pub async fn execute(&self, client: Client) -> Result<(), CliError> {
+    pub async fn execute<AUTH>(&self, client: Client<AUTH>) -> Result<(), CliError> {
         match self {
             TagsCmd { add: Some(tag), .. } => {
                 add_tag(client, *tag).await?;
@@ -41,7 +41,7 @@ impl TagsCmd {
 
 // HELPERS
 // ================================================================================================
-async fn list_tags(client: Client) -> Result<(), CliError> {
+async fn list_tags<AUTH>(client: Client<AUTH>) -> Result<(), CliError> {
     let (cli_config, _) = load_config_file()?;
     let mut table = create_dynamic_table(&["Tag", "Source"]);
 
@@ -65,7 +65,7 @@ async fn list_tags(client: Client) -> Result<(), CliError> {
     Ok(())
 }
 
-async fn add_tag(mut client: Client, tag: u32) -> Result<(), CliError> {
+async fn add_tag<AUTH>(mut client: Client<AUTH>, tag: u32) -> Result<(), CliError> {
     let tag: NoteTag = tag.into();
     let execution_mode = match tag.execution_mode() {
         NoteExecutionMode::Local => "Local",
@@ -81,7 +81,7 @@ async fn add_tag(mut client: Client, tag: u32) -> Result<(), CliError> {
     Ok(())
 }
 
-async fn remove_tag(mut client: Client, tag: u32) -> Result<(), CliError> {
+async fn remove_tag<AUTH>(mut client: Client<AUTH>, tag: u32) -> Result<(), CliError> {
     client.remove_note_tag(tag.into()).await?;
     println!("Tag {tag} removed");
     Ok(())