+++ bin/miden-cli/tests/cli.rs
@@ -21,7 +21,8 @@ use miden_client::{
     testing::{
         account_id::ACCOUNT_ID_PRIVATE_SENDER,
         common::{
-            ACCOUNT_ID_REGULAR, TEST_CLIENT_RPC_CONFIG_FILE, execute_tx_and_sync, insert_new_wallet,
+            ACCOUNT_ID_REGULAR, TEST_CLIENT_RPC_CONFIG_FILE, TestClientKeyStore,
+            execute_tx_and_sync, insert_new_wallet,
         },
     },
     transaction::{OutputNote, TransactionRequestBuilder},
@@ -737,7 +738,7 @@ pub fn create_test_store_path() -> std::path::PathBuf {
     temp_file
 }
 
-pub type TestClient = Client;
+pub type TestClient = Client<TestClientKeyStore>;
 
 /// Creates a new [`Client`] with a given store. Also returns the keystore associated with it.
 async fn create_rust_client_with_store_path(store_path: &Path) -> (TestClient, CliKeyStore) {