+++ crates/web-client/test/fpi.test.ts
@@ -66,18 +66,7 @@ export const testStandardFpi = async (): Promise<void> => {
     );
     await client.syncState();
 
-    let deploymentTxScript = window.TransactionScript.compile(
-      `
-                begin 
-                    call.::miden::contracts::auth::basic::auth__tx_rpo_falcon512 
-                end
-            `,
-      window.TransactionKernel.assembler()
-    );
-
-    let txRequest = new window.TransactionRequestBuilder()
-      .withCustomScript(deploymentTxScript)
-      .build();
+    let txRequest = new window.TransactionRequestBuilder().build();
 
     let txResult = await client.newTransaction(foreignAccountId, txRequest);
 
@@ -108,8 +97,6 @@ export const testStandardFpi = async (): Promise<void> => {
         
                 exec.tx::execute_foreign_procedure
                 push.9.12.18.30 assert_eqw
-        
-                call.::miden::contracts::auth::basic::auth__tx_rpo_falcon512 
             end
         `;
     txScript = txScript