+++ crates/web-client/test/new_transactions.test.ts
@@ -300,6 +300,8 @@ describe("swap transaction tests", () => {
         assetAAmount,
         faucetB,
         assetBAmount,
+        "private",
+        "private",
         flag
       );
 
@@ -554,7 +556,6 @@ export const customTransaction = async (
       // Just like in the miden test, you can modify this script to get the execution to fail
       // by modifying the assert
       let txScript = `
-            use.miden::contracts::auth::basic->auth_tx
             use.miden::kernels::tx::prologue
             use.miden::kernels::tx::memory
 
@@ -562,8 +563,6 @@ export const customTransaction = async (
                 push.0 push.${_assertedValue}
                 # => [0, ${_assertedValue}]
                 assert_eq
-
-                call.auth_tx::auth__tx_rpo_falcon512
             end
         `;
 
@@ -1099,7 +1098,8 @@ export const counterAccountComponent = async (): Promise<
       `;
     const incrNonceAuthCode = `use.miden::account
         export.auth__basic
-          push.1 exec.account::incr_nonce
+          exec.account::incr_nonce
+          drop
         end`;
     const client = window.client;